+++ lib/Screens/Reading/Widgets/group_reading_card.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class GroupReadingCard extends StatelessWidget {
+  const GroupReadingCard({
+    required this.image,
+    required this.title,
+    required this.press,
+  });
+
+  final String image, title;
+  final Function() press;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: press,
+      child: Card(
+        elevation: 1,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(10.0),
+        ),
+        clipBehavior: Clip.antiAlias,
+        child: Column(
+          children: [
+            ListTile(
+              title: Text(this.title),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
