+++ lib/Screens/Reading/group_reading_cards.dart
@@ -1,73 +1,42 @@
-import 'package:app_mental/Screens/Reading/group_list_screen.dart';
+import 'package:app_mental/Screens/Reading/Widgets/group_reading_card.dart';
+import 'package:app_mental/Screens/Reading/intervention_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class GroupReadingCards extends StatefulWidget {
-  final List<String> groupList;
+class GroupReadingCardsList extends StatefulWidget {
+  final List<String> readingGroupList;
 
-  GroupReadingCards({required this.groupList});
+  GroupReadingCardsList({required this.readingGroupList});
 
   @override
-  State<GroupReadingCards> createState() => _GroupReadingCardsState();
+  State<GroupReadingCardsList> createState() => _GroupReadingCardsListState();
 }
 
-class _GroupReadingCardsState extends State<GroupReadingCards> {
+class _GroupReadingCardsListState extends State<GroupReadingCardsList> {
   @override
   Widget build(BuildContext context) {
     return Expanded(
       child: ListView.builder(
         padding: EdgeInsets.all(10.0),
         scrollDirection: Axis.vertical,
-        itemCount: widget.groupList.length,
+        itemCount: widget.readingGroupList.length,
         itemBuilder: (BuildContext context, int index) {
-          var group = widget.groupList.elementAt(index);
+          var readingGroup = widget.readingGroupList.elementAt(index);
           return GroupReadingCard(
             image: "assets/images/reading01.jpg",
-            title: group,
+            title: readingGroup,
             press: () {
               Navigator.push(
                 context,
                 MaterialPageRoute(
                   builder: (context) {
-                    return GroupListScreen(group);
+                    return InterventionScreen(readingGroup);
                   },
                 ),
               );
             },
           );
         },
-      ),
-    );
-  }
-}
-
-class GroupReadingCard extends StatelessWidget {
-  const GroupReadingCard({
-    required this.image,
-    required this.title,
-    required this.press,
-  });
-
-  final String image, title;
-  final Function() press;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: press,
-      child: Card(
-        elevation: 1,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10.0),
-        ),
-        clipBehavior: Clip.antiAlias,
-        child: Column(
-          children: [
-            ListTile(
-              title: Text(this.title),
-            ),
-          ],
-        ),
       ),
     );
   }
